🔀 :: 인증번호 검증 기능 추가

+++ account/src/main/kotlin/com/project/indistraw/account/application/port/output/dto/TokenDto.kt
@@ -3,6 +3,6 @@ package com.project.indistraw.account.application.port.output.dto
 data class TokenDto(
     val accessToken: String,
     val refreshToken: String,
-    val accessTokenExpiredAt: Int,
-    val refreshTokenExpiredAt: Int
+    val accessTokenExpiredAt: Long,
+    val refreshTokenExpiredAt: Long
 )
