+++ account/src/main/kotlin/com/project/indistraw/account/domain/RefreshToken.kt
@@ -5,5 +5,5 @@ import java.util.UUID
 data class RefreshToken(
     val refreshToken: String,
     val accountIdx: UUID,
-    val expiredAt: Int
+    val expiredAt: Long
 )
